refactor(gtdbtk.io): Refactored package to be files instead of io.
