+++ gtdbtk/files/batchfile.py
