+++ gtdbtk/files/marker/tophit.py
