+++ gtdbtk/files/marker_info.py
