+++ gtdbtk/files/red_dict.py
