+++ gtdbtk/files/tree_mapping.py
